requirements.txt
annotated-types==0.5.0
cachetools==5.3.1
fastapi==0.101.1
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
numpy==1.25.2
pandas==2.0.3
proto-plus==1.22.3
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.2.1
pydantic_core==2.6.1
pytz==2023.3
rsa==4.9
starlette==0.27.0
tzdata==2023.3
urllib3==1.26.16
